rename to rabbit-ng
