+++ src/rabbit_ng/errors.py
