+++ src/rabbit_ng/resources/models/bimbas.joblib
