+++ tests/test_main.py
@@ -2,20 +2,20 @@ from unittest.mock import patch
 
 import pytest
 
-from rabbit.main import (
+from rabbit_ng.main import (
     _process_single_contributor,
     run_rabbit,
 )
-from rabbit.errors import RabbitErrors
-from rabbit.predictor import ContributorResult
+from rabbit_ng.errors import RabbitErrors
+from rabbit_ng.predictor import ContributorResult
 
 
 class TestProcessSingleContributor:
     """Tests for the _process_single_contributor function."""
 
-    @patch("rabbit.main.ONNXPredictor")
-    @patch("rabbit.main.GitHubAPIExtractor")
-    @patch("rabbit.main.predict_user_type")
+    @patch("rabbit_ng.main.ONNXPredictor")
+    @patch("rabbit_ng.main.GitHubAPIExtractor")
+    @patch("rabbit_ng.main.predict_user_type")
     def test_process_contributor_success(
         self, mock_predict, mock_gh_extractor, mock_predictor
     ):
@@ -39,8 +39,8 @@ class TestProcessSingleContributor:
         assert result.user_type == "Human"
         assert result.confidence == 0.95
 
-    @patch("rabbit.main.ONNXPredictor")
-    @patch("rabbit.main.GitHubAPIExtractor")
+    @patch("rabbit_ng.main.ONNXPredictor")
+    @patch("rabbit_ng.main.GitHubAPIExtractor")
     def test_process_contributor_organisation(self, mock_gh_extractor, mock_predictor):
         """Test if _process_single_contributor return directly when user is an Organisation."""
         mock_gh_extractor.query_user_type.return_value = "Organization"
@@ -57,8 +57,8 @@ class TestProcessSingleContributor:
         assert result.user_type == "Organization"
         assert result.confidence == 1.0
 
-    @patch("rabbit.main.ONNXPredictor")
-    @patch("rabbit.main.GitHubAPIExtractor")
+    @patch("rabbit_ng.main.ONNXPredictor")
+    @patch("rabbit_ng.main.GitHubAPIExtractor")
     def test_process_contributor_app(self, mock_gh_extractor, mock_predictor):
         """Test if _process_single_contributor return directly when user is an Organisation."""
         mock_gh_extractor.query_user_type.return_value = "Bot"
@@ -75,9 +75,9 @@ class TestProcessSingleContributor:
         assert result.user_type == "Bot"
         assert result.confidence == 1.0
 
-    @patch("rabbit.main.ONNXPredictor")
-    @patch("rabbit.main.GitHubAPIExtractor")
-    @patch("rabbit.main.predict_user_type")
+    @patch("rabbit_ng.main.ONNXPredictor")
+    @patch("rabbit_ng.main.GitHubAPIExtractor")
+    @patch("rabbit_ng.main.predict_user_type")
     def test_process_contributor_with_more_than_100_event_minimum(
         self, mock_predict, mock_gh_extractor, mock_predictor
     ):
@@ -106,8 +106,8 @@ class TestProcessSingleContributor:
         assert result.user_type == "Bot"
         assert result.confidence == 0.90
 
-    @patch("rabbit.main.ONNXPredictor")
-    @patch("rabbit.main.GitHubAPIExtractor")
+    @patch("rabbit_ng.main.ONNXPredictor")
+    @patch("rabbit_ng.main.GitHubAPIExtractor")
     def test_process_contributor_returns_unknown_when_no_events(
         self, mock_gh_extractor, mock_predictor
     ):
@@ -130,13 +130,13 @@ class TestProcessSingleContributor:
         assert result.user_type == "Unknown"
         assert result.confidence == "-"
 
-    @patch("rabbit.main.ONNXPredictor")
-    @patch("rabbit.main.GitHubAPIExtractor")
+    @patch("rabbit_ng.main.ONNXPredictor")
+    @patch("rabbit_ng.main.GitHubAPIExtractor")
     def test_process_contributor_returns_invalid_when_not_found(
         self, mock_gh_extractor, mock_predictor
     ):
         """Test _process_single_contributor handles NotFoundError correctly."""
-        from rabbit.errors import NotFoundError
+        from rabbit_ng.errors import NotFoundError
 
         mock_gh_extractor.query_user_type.side_effect = NotFoundError("User not found")
 
@@ -152,9 +152,9 @@ class TestProcessSingleContributor:
         assert result.user_type == "Invalid"
         assert result.confidence == "-"
 
-    @patch("rabbit.main.ONNXPredictor")
-    @patch("rabbit.main.GitHubAPIExtractor")
-    @patch("rabbit.main.predict_user_type")
+    @patch("rabbit_ng.main.ONNXPredictor")
+    @patch("rabbit_ng.main.GitHubAPIExtractor")
+    @patch("rabbit_ng.main.predict_user_type")
     def test_process_contributor_with_early_stopping(
         self, mock_predict, mock_gh_extractor, mock_predictor
     ):
@@ -188,13 +188,13 @@ class TestProcessSingleContributor:
         # Make sure that it was called only once
         assert mock_gh_extractor.query_events.call_count == 1
 
-    @patch("rabbit.main.ONNXPredictor")
-    @patch("rabbit.main.GitHubAPIExtractor")
+    @patch("rabbit_ng.main.ONNXPredictor")
+    @patch("rabbit_ng.main.GitHubAPIExtractor")
     def test_process_contributor_forwards_api_errors(
         self, mock_gh_extractor, mock_predictor
     ):
         """Test _process_single_contributor forwards GitHubAPIError exceptions."""
-        from rabbit.errors import RabbitErrors
+        from rabbit_ng.errors import RabbitErrors
 
         mock_gh_extractor.query_events.side_effect = RabbitErrors("API error")
         mock_gh_extractor.query_user_type.return_value = "User"
@@ -208,8 +208,8 @@ class TestProcessSingleContributor:
                 min_confidence=1,
             )
 
-    @patch("rabbit.main.ONNXPredictor")
-    @patch("rabbit.main.GitHubAPIExtractor")
+    @patch("rabbit_ng.main.ONNXPredictor")
+    @patch("rabbit_ng.main.GitHubAPIExtractor")
     def test_process_contributor_forwards_unexpected_errors(
         self, mock_gh_extractor, mock_predictor
     ):
@@ -231,7 +231,7 @@ class TestProcessSingleContributor:
 class TestRunRabbit:
     """Tests for the run_rabbit function."""
 
-    @patch("rabbit.main._process_single_contributor")
+    @patch("rabbit_ng.main._process_single_contributor")
     def test_run_rabbit_multiple_contributors(self, mock_process):
         """Test run_rabbit processes multiple contributors correctly."""
         sample_result = ContributorResult(
